+++ other/upgrade_2-1_sqlite.sql
@@ -442,7 +442,7 @@ CREATE TABLE IF NOT EXISTS {$db_prefix}user_alerts (
   content_type varchar(255) NOT NULL default '',
   content_id int unsigned NOT NULL default '0',
   content_action varchar(255) NOT NULL default '',
-  is_read smallint unsigned NOT NULL default '0',
+  is_read int unsigned NOT NULL default '0',
   extra text NOT NULL,
   PRIMARY KEY (id_alert)
 );
